User experience improvements
Updated the Beer rating logic so it renders the “thank you” message
immediately after rating, instead of waiting for a server response.
Also, fixed the rendering of Share buttons when sharing Pages and Posts
and fixed an issue where navigating backwards in history while a share
modal was open would break scrolling.