Improved user experience on interior page navigation
For pages other than the home page, navigation now loads more
immediately to the page content by adding some clever
non-render-blocking scripts. It can be further improved, but this is a
step in the right direction as it was previously waiting until the
entire header loaded before displaying the page content.