Improved Firefox experience
Fixed a bug in Firefox where navigating between pages would cause the
content to shift to the left due to the way the browser handles
flexbox’s `justify-content` property with absolutely positioned
children. See [this StackOverflow
post](http://stackoverflow.com/questions/28956197/flexbox-style-justify-
content-space-between-misaligned-in-firefox-with-absolu) for additional
information.

Moved loading indicator out of the `main` content wrapper in order
eliminate it’s impact on `space-between`.