IE Edge FTW
Can’t use `calc` anywhere with transitions if you want to support Edge
because who knows [when they will fix the
bug](https://connect.microsoft.com/IE/feedback/details/762719/css3-calc-
bug-inside-transition-or-transform). LOL @ “based on the limited impact
this bug may have, we will not be able to address this bug during this
release. “ A bug is a bug, M$!