Minor bug fixes
Fixed an issue where desktop browsers would not show the content after
navigating to the Beer list, then going to a beer page, for the first
time the page is loaded.

Fixed an issue where the wrong Business Hours was being highlighted
after omitting the Closed days from the list.